Fix the bug where an error was handled as `AggregateException` rather than the original error in Connections Clean up some unreached code
