+++ Neo4j.Driver/Neo4j.Driver.IntegrationTests/ExamplesAsync.cs
@@ -856,7 +856,6 @@ namespace Neo4j.Driver.ExamplesAsync
             {
                 var result = session.Run("MATCH (n) DETACH DELETE n");
                 result.Consume();
-
             }
         }
